bdev: clean up init_read and init_write apis

+++ include/spdk/scsi.h
@@ -130,7 +130,7 @@ struct spdk_scsi_task {
 	 */
 	struct iovec iov;
 	struct iovec *iovs;
-	uint16_t iovcnt;
+	int iovcnt;
 
 	uint8_t sense_data[32];
 	size_t sense_data_len;
